UPSTREAM: init, tracing: Add initcall trace events
Being able to trace the start and stop of initcalls is useful to see where
the timings are an issue. There is already an "initcall_debug" parameter,
but that can cause a large overhead itself, as the printing of the
information may take longer than the initcall functions.

Adding in a start and finish trace event around the initcall functions, as
well as a trace event that records the level of the initcalls, one can get a
much finer measurement of the times and interactions of the initcalls
themselves, as trace events are much lighter than printk()s.

Suggested-by: Abderrahmane Benbachir <abderrahmane.benbachir@polymtl.ca>
Change-Id: I18524f8aa48a9361ddf780854db20e4b933bbd29
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: slipzSIX <slipzsix@gmail.com>