treewide: Disable CPU wakeup log
This kernel will explicitly, and always be used in Production builds. It is not needed to keep the logging enabled
each time a CPU core is taken offline and/or sent to deep sleep or woken up. It causes more log processes
to trigger, leading to increased power consumption as well as CPU overhead.

Change-Id: I4120ddad00bb0650934953819f987f06c4848899
Signed-off-by: Kanishk <kanishkthederp@gmail.com>
Signed-off-by: Niranjan BR <niranjankannan2003@gmail.com>
Signed-off-by: Frost <frost.github@proton.me>
Signed-off-by: slipzSIX <slipzsix@gmail.com>